Fix search, add searching for alternative titles
The search terms were being encoded twice, resulting in '%2B' instead of
'+', etc. This was not obvious with spaces, but broke search results for
other characters.

We now look up alternative titles on TheTVDB.com (since the 'titles' list
is always empty) and also search for those. This is often necessary when
the English title of a series is different from the original title.